feat(textures): adopt core's TextureId, key the registry on it, add nativeTextureId (#330)
Phase 2 of labelle-gfx#328 (RFC-TEXTURE-ID-TYPING). Pins labelle-core
v1.28.0 for the types.

- `TextureId` is now core's type rather than a gfx-local duplicate, so
  gfx, the backends, the engine and games converge on ONE nominal type.
  `BackendTextureId` is re-exported alongside it.
- The registry is keyed on `TextureId` instead of a bare `u32`. This is
  what makes the guarantee structural: a BACKEND id is a different type
  and can no longer be used as a key by accident.
- New `nativeTextureId(TextureId) ?BackendTextureId` — the one legal
  conversion between the id spaces, on the engine and forwarded through
  `GfxRenderer`. Before it existed, callers handed a `TextureId` straight
  to a backend accessor and it compiled, which is how #326 silently
  blanked a downstream menu.

Core ships no `from(u32)` on purpose, so the ~11 `TextureId.from(...)`
call sites became explicit `@enumFromInt`. Two remain in src and both
are engine-facing seams (`drawMesh`, `drawScreenTexture`) that still
take a bare `u32`; typing those is phase 4. `mintTextureKey` keeps the
only allocation-site cast.

`nativeTextureId` currently retags `info.backend_texture.id` because
backends still declare `Texture.id: u32`. Marked SCAFFOLDING — phase 3
gives them `BackendTextureId` and the retag goes away.

Tests exercise the seam through BOTH the engine and `GfxRenderer` — a
forwarder that exists only on the inner type is a no-op in practice
(the gfx#291 lesson). Verified non-vacuous: breaking the wrapper
forwarder fails the suite. 147/147.

+++ src/retained_engine/draw.zig
@@ -30,7 +30,7 @@ pub fn DrawHelpers(comptime Self: type) type {
         pub fn drawSpriteEntry(self: *const Self, entry: *const SpriteEntry) void {
             const sprite = &entry.visual;
             const pos = entry.position;
-            const tex_id = sprite.texture.toInt();
+            const tex_id = sprite.texture;
 
             // Resolve source rect and display dimensions
             const tex_info = self.textures.get(tex_id);
@@ -80,9 +80,9 @@ pub fn DrawHelpers(comptime Self: type) type {
             // texture has been unloaded, so fabricating from one can
             // only ever produce garbage. Draw nothing instead
             // (labelle-gfx#324).
-            if (tex_info == null and tex_id >= Self.TEXTURE_KEY_BASE) return;
+            if (tex_info == null and tex_id.toInt() >= Self.TEXTURE_KEY_BASE) return;
             const backend_tex: B.Texture = if (tex_info) |t| t.backend_texture else .{
-                .id = tex_id,
+                .id = tex_id.toInt(),
                 .width = @intFromFloat(display_w),
                 .height = @intFromFloat(display_h),
             };
